After review 1

+++ api_yamdb/reviews/models.py
@@ -1,6 +1,6 @@
-
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
+
 from users.models import User
 
 from .validators import validate_year
@@ -122,7 +122,7 @@ class Review(models.Model):
         related_name='reviews',
         verbose_name='Автор отзыва'
     )
-    score = models.IntegerField(
+    score = models.PositiveSmallIntegerField(
         default=1,
         blank=True,
         verbose_name='Оценка',
